@@ -1,6 +1,7 @@
 {
   "name": "my-vue-app",
   "version": "0.1.0",
+   "type": "module",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -13,16 +14,19 @@
     "@fortawesome/free-regular-svg-icons": "^6.6.0",
     "@fortawesome/free-solid-svg-icons": "^6.6.0",
     "@fortawesome/vue-fontawesome": "^3.0.8",
+    "@vue/server-renderer": "^3.5.11",
     "@vueuse/head": "^2.0.0",
     "@vueuse/motion": "^2.2.5",
     "axios": "^1.7.7",
     "core-js": "^3.8.3",
     "date-fns": "^4.1.0",
+    "express": "^4.21.0",
     "lodash": "^4.17.21",
     "mitt": "^3.0.1",
+    "quill": "^2.0.2",
     "simplebar": "^6.2.7",
     "toastify-js": "^1.12.0",
-    "vue": "^3.2.13",
+    "vue": "^3.5.11",
     "vue-icons": "^2.0.0",
     "vue-infinite-scroll-component": "^1.0.3",
     "vue-router": "^4.4.5",
