@@ -1,21 +1,31 @@
 <template>
-  <div>
+  <div class="bookmarks-container">
     <h1>Your Bookmarks</h1>
-    <div v-if="loading">Loading bookmarks...</div>
+    <div v-if="loading" class="loading-message">Loading bookmarks...</div>
     <div v-else>
-      <ul>
-        <li v-for="bookmark in bookmarks" :key="bookmark.id" class="bookmark-item">
-          <img :src="bookmark.full_image_url" alt="Post Image" class="bookmark-image" />
-          <div class="bookmark-details">
-            <h3>{{ bookmark.title }}</h3>
-            <button @click="removeBookmark(bookmark.id)">Remove</button>
+      <div class="grid">
+        <div v-for="bookmark in bookmarks" :key="bookmark.id" class="bookmark-card">
+          <div class="image-container">
+            <router-link :to="`/${bookmark.category_slug}/${bookmark.slug}`" class="bookmark-link">
+              <img :src="bookmark.full_image_url" alt="Post Image" class="bookmark-image" />
+            </router-link>
+            <i @click="removeBookmark(bookmark.id)" class="fas fa-trash remove-icon"></i>
           </div>
-        </li>
-      </ul>
-      <div v-if="bookmarks.length === 0">No bookmarks found.</div>
+          <div class="bookmark-details">
+            <h3>
+              <router-link :to="`/${bookmark.category_slug}/${bookmark.slug}`" class="bookmark-link">
+                {{ truncateTitle(bookmark.title, 30) }} <!-- Limit to 30 characters -->
+              </router-link>
+            </h3>
+          </div>
+        </div>
+      </div>
+      <div v-if="bookmarks.length === 0" class="no-bookmarks">No bookmarks found.</div>
     </div>
   </div>
 </template>
+
+
 
 <script>
 import axios from 'axios';
@@ -25,14 +35,22 @@ export default {
   setup() {
     const bookmarks = ref([]);
     const loading = ref(true);
-    const userId = 1; // Replace this with the actual user ID
+
+    const loggedInUser = localStorage.getItem('user') || sessionStorage.getItem('user');
+    const userId = loggedInUser ? JSON.parse(loggedInUser).id : null;
 
     const fetchBookmarks = async () => {
+      if (!userId) {
+        console.error('No user ID found');
+        loading.value = false;
+        return;
+      }
+
       try {
         const response = await axios.get('https://blog.tourismofkashmir.com/bookmark_view_api.php', {
           params: { user_id: userId },
         });
-        bookmarks.value = response.data; // Ensure the response includes an imageUrl field
+        bookmarks.value = response.data; // Ensure the response includes the required fields
       } catch (error) {
         console.error('Error fetching bookmarks:', error);
       } finally {
@@ -41,6 +59,8 @@ export default {
     };
 
     const removeBookmark = async (bookmarkId) => {
+      if (!userId) return;
+
       try {
         await axios.delete('https://blog.tourismofkashmir.com/bookmark_view_api.php', {
           params: { user_id: userId, action: 'delete', post_id: bookmarkId },
@@ -51,35 +71,112 @@ export default {
       }
     };
 
+    const truncateTitle = (title, maxLength) => {
+      if (title.length > maxLength) {
+        return title.substring(0, maxLength) + '...';
+      }
+      return title;
+    };
+
     onMounted(fetchBookmarks);
 
     return {
       bookmarks,
       loading,
       removeBookmark,
+      truncateTitle,
     };
   },
 };
 </script>
 
+
 <style scoped>
-.bookmark-item {
-  display: flex;
-  align-items: center;
+.bookmarks-container {
+  padding: 20px;
+  max-width: 800px;
+  margin: auto;
+  font-family: 'Arial', sans-serif;
+}
+
+h1 {
+  text-align: center;
   margin-bottom: 20px;
+  font-size: 24px;
+  color: #333;
+}
+
+.loading-message {
+  text-align: center;
+  font-size: 18px;
+  color: #999;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+  gap: 15px;
+}
+
+.bookmark-card {
+  background-color: #fff;
+  border-radius: 12px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+  transition: transform 0.2s;
+}
+
+.bookmark-card:hover {
+  transform: translateY(-5px);
+}
+
+.image-container {
+  position: relative; /* Enables absolute positioning for child elements */
 }
 
 .bookmark-image {
-  width: 100px; /* Adjust size as needed */
+  width: 100%;
   height: auto;
-  margin-right: 20px; /* Space between image and title */
+}
+
+.remove-icon {
+  color: black;
+  font-size: 18px;
+  cursor: pointer;
+  position: absolute;
+  right: 10px; /* Adjust for positioning */
+  top: 10px;  /* Adjust for positioning */
+  transition: color 0.2s;
+}
+
+.remove-icon:hover {
+  color: #c0392b;
 }
 
 .bookmark-details {
-  flex-grow: 1; /* Allow details to take available space */
+  padding: 10px;
 }
 
 .bookmark-details h3 {
-  margin: 0;
+  margin: 10px 0;
+  font-size: 18px;
+  color: #333;
+}
+
+.bookmark-link {
+  text-decoration: none;
+  color: #2e86de;
+  transition: color 0.2s;
+}
+
+.bookmark-link:hover {
+  color: #1e60b0;
+}
+
+.no-bookmarks {
+  text-align: center;
+  color: #999;
+  font-size: 18px;
+  margin-top: 20px;
 }
 </style>
