@@ -6,23 +6,22 @@
         <span class="custom-close-button" @click="onClose">&times;</span>
       </div>
       <div class="modal-body" @scroll="handleScroll" ref="modalBody">
-        
-    <div v-if="initialLoading" class="aly-loading-placeholder">
-  <div class="aly-loading-comment" v-for="index in 7" :key="index">
-    <div class="aly-loading-avatar"></div>
-    <div class="aly-loading-details">
-      <div class="aly-loading-line aly-loading-name"></div>
-      <div class="aly-loading-line aly-loading-message"></div>
-    </div>
-  </div>
-</div>
+        <div v-if="initialLoading" class="aly-loading-placeholder">
+          <div class="aly-loading-comment" v-for="index in 7" :key="index">
+            <div class="aly-loading-avatar"></div>
+            <div class="aly-loading-details">
+              <div class="aly-loading-line aly-loading-name"></div>
+              <div class="aly-loading-line aly-loading-message"></div>
+            </div>
+          </div>
+        </div>
 
         <div v-else>
           <div v-if="comments.length === 0" class="no-comments-message">No comments yet.</div>
           <div v-else>
             <div v-for="(comment, index) in comments" :key="comment.id" class="custom-comment-item">
               <div class="custom-comment-avatar">
-                <img :src="comment.avatar" alt="Avatar"/>
+                <img :src="comment.avatar" alt="Avatar" />
               </div>
               <div class="custom-comment-content">
                 <div class="custom-comment-header">
@@ -39,7 +38,7 @@
                     <a href="#!" class="custom-read-more-link" @click="toggleFullComment(index)">Read less</a>
                   </span>
                   <span v-else>
-                    {{ comment.content.length > 100 ? `${comment.content.substring(0, 100)}...` : comment.content }}
+                    {{ comment.content.length > 100 ? comment.content.substring(0, 100) + '...' : comment.content }}
                     <a v-if="comment.content.length > 100" href="#!" class="custom-read-more-link" @click="toggleFullComment(index)">Read more</a>
                   </span>
                 </div>
@@ -54,27 +53,32 @@
       </div>
 
       <div class="comment-input-container">
-  <input 
-    v-model="newComment" 
-    placeholder="Write a comment..." 
-    class="custom-comment-input" 
-    type="text" />
-  <button @click="postComment" class="custom-post-button">Post</button>
-</div>
-
+        <input 
+          v-model="newComment" 
+          placeholder="Write a comment..." 
+          class="custom-comment-input" 
+          type="text" 
+        />
+        <button 
+          @click="postComment" 
+          class="custom-post-button" 
+          :disabled="!newComment.trim()"
+        >
+          Post
+        </button>
+      </div>
     </div>
   </div>
 </template>
 
-
-
-
 <script>
-import { ref, onMounted } from 'vue';
+import { ref, onMounted, onBeforeUnmount } from 'vue';
 import axios from 'axios';
 import { formatDistanceToNow } from 'date-fns';
 import './comment.css';
 import _ from 'lodash';
+import Toastify from 'toastify-js'; // Import Toastify
+import 'toastify-js/src/toastify.css'; // Import Toastify CSS
 
 export default {
   props: {
@@ -123,45 +127,51 @@ export default {
     };
 
     const postComment = async () => {
-  const loggedInUser = localStorage.getItem('user');
-  if (!loggedInUser) {
-    alert("Please log in to post a comment");
-    return;
-  }
+      const loggedInUser = localStorage.getItem('user');
+      if (!loggedInUser) {
+        Toastify({
+          text: "Please log in to post a comment",
+          duration: 3000,
+          backgroundColor: "linear-gradient(to right, #ff5f6d, #ffc371)",
+          className: "info",
+        }).showToast();
+        return;
+      }
 
-  if (!newComment.value.trim()) return;
+      if (!newComment.value.trim()) return;
 
-  const user = JSON.parse(loggedInUser);
-  const commentData = {
-    post_id: props.postId,
-    user_id: user.id,
-    content: newComment.value,
-    created_at: new Date().toISOString(), // Add the current timestamp
-  };
+      const user = JSON.parse(loggedInUser);
+      const commentData = {
+        post_id: props.postId,
+        user_id: user.id,
+        content: newComment.value,
+        created_at: new Date().toISOString(), // Add the current timestamp
+      };
 
-  try {
-    const response = await axios.post('https://blog.tourismofkashmir.com/api_comments.php', commentData);
-    
-    // Prepend the new comment to the comments array
-    comments.value.unshift({
-      id: response.data.id, // Assuming your API returns the new comment's ID
-      avatar: user.avatar, // Add user's avatar or a default one
-      username: user.username, // Add user's username
-      content: newComment.value,
-      created_at: commentData.created_at, // Use the created_at timestamp
-    });
-    
-    newComment.value = ''; // Clear the input
-  } catch (error) {
-    console.error("Error posting comment:", error);
-  }
-};
+      try {
+        const response = await axios.post('https://blog.tourismofkashmir.com/api_comments.php', commentData);
+        
+        // Prepend the new comment to the comments array
+        comments.value.unshift({
+          id: response.data.id, // Assuming your API returns the new comment's ID
+          avatar: user.avatar, // Add user's avatar or a default one
+          username: user.username, // Add user's username
+          content: newComment.value,
+          created_at: commentData.created_at, // Use the created_at timestamp
+        });
+        
+        newComment.value = ''; // Clear the input
+      } catch (error) {
+        console.error("Error posting comment:", error);
+      }
+    };
 
     const toggleFullComment = (index) => {
       showFullComment.value[index] = !showFullComment.value[index];
     };
 
     const handleScroll = _.throttle(() => {
+      if (!modalBody.value) return; // Check if modalBody is available
       const scrollTop = modalBody.value.scrollTop;
       const scrollHeight = modalBody.value.scrollHeight;
       const clientHeight = modalBody.value.clientHeight;
@@ -173,6 +183,15 @@ export default {
 
     onMounted(() => {
       fetchComments();
+      if (modalBody.value) {
+        modalBody.value.addEventListener('scroll', handleScroll);
+      }
+    });
+
+    onBeforeUnmount(() => {
+      if (modalBody.value) {
+        modalBody.value.removeEventListener('scroll', handleScroll);
+      }
     });
 
     return {
@@ -190,7 +209,4 @@ export default {
   },
 };
 </script>
-
-
-
 
