@@ -1,11 +1,22 @@
-import { createRouter, createWebHistory } from 'vue-router';
-import NewsList from '../components/NewsList.vue'; // Adjust the path based on your folder structure
-import PostDetail from '../components/NewsDetail.vue'; // Adjust the path based on your folder structure
-import CategoryList from '../components/CategoryList.vue'; // Import the CategoryList component
-import LoginPage from '../components/LoginPage.vue';
-import NotificationPage from '../components/NotificationPage.vue'; // Import the NotificationPage component
-import BookmarkPage from '../components/BookmarkPage.vue'; // Import the BookmarkPage component
+// src/router/index.js
+const { createRouter, createWebHistory } = require('vue-router');
 
+// Importing components
+const NewsList = require('../components/NewsList.vue');
+const PostDetail = require('../components/NewsDetail.vue');
+const CategoryList = require('../components/CategoryList.vue');
+const LoginPage = require('../components/LoginPage.vue');
+const NotificationPage = require('../components/NotificationPage.vue');
+const BookmarkPage = require('../components/BookmarkPage.vue');
+const ProfilePage = require('../components/ProfilePage.vue');
+const PageViewer = require('../components/PageViewer.vue');
+const TagDetails = require('../components/TagDetails.vue');
+const DashboardPage = require('../Dashboard/DashboardPage.vue');
+const CreatePost = require('../Dashboard/CreatePost.vue');
+const PostList = require('../Dashboard/PostList.vue');
+const EditPost = require('../Dashboard/EditPost.vue');
+
+// Define routes
 const routes = [
   {
     path: '/',
@@ -14,13 +25,14 @@ const routes = [
   },
   {
     path: '/:categorySlug',
+    name: 'Category',
     component: CategoryList,
   },
   {
     path: '/:category_slug/:slug',
     name: 'PostDetail',
     component: PostDetail,
-    props: true, // Pass route params as props to the component
+    props: true,
   },
   {
     path: '/login',
@@ -29,19 +41,59 @@ const routes = [
   },
   {
     path: '/notification',
-    name: 'notifications', // Ensure this name matches the computed property check
-    component: NotificationPage, // Add the new notification route
+    name: 'Notifications',
+    component: NotificationPage,
   },
   {
     path: '/bookmark',
     name: 'Bookmark',
-    component: BookmarkPage, // Add the new bookmark route
+    component: BookmarkPage,
+  },
+  {
+    path: '/profile/:username',
+    name: 'Profile',
+    component: ProfilePage,
+    props: true,
+  },
+  {
+    path: '/page/:slug',
+    name: 'PageViewer',
+    component: PageViewer,
+  },
+  {
+    path: '/tags/:tagSlug',
+    name: 'TagDetails',
+    component: TagDetails,
+    props: true,
+  },
+  {
+    path: '/dashboard',
+    name: 'Dashboard',
+    component: DashboardPage,
+  },
+  {
+    path: '/dashboard/create-post',
+    name: 'CreatePost',
+    component: CreatePost,
+  },
+  {
+    path: '/dashboard/post-list',
+    name: 'PostList',
+    component: PostList,
+  },
+  {
+    path: '/dashboard/edit-post/:id',
+    name: 'EditPost',
+    component: EditPost,
+    props: true,
   },
 ];
 
+// Create router instance
 const router = createRouter({
-  history: createWebHistory(process.env.BASE_URL),
+  history: createWebHistory(),
   routes,
 });
 
-export default router;
+// Export the router instance
+module.exports = router; // Use CommonJS export
